% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibw2992.R
\docType{data}
\name{bibw2992}
\alias{bibw2992}
\title{The 'bibw2992' Dataset}
\format{A data frame with N rows and P columns}
\usage{
bibw2992
}
\description{
bibw2992 is a dataset used to power the examples run in this package
}
\keyword{datasets}
