% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNF_CPLEX_weak_pos.R
\name{CNF_CPLEX_weak_pos}
\alias{CNF_CPLEX_weak_pos}
\title{CNF CPLEX weak position function}
\usage{
CNF_CPLEX_weak_pos(X, Y, W, K, M, lambda, sens, spec, addcons)
}
\arguments{
\item{X}{An N x P binary matrix with N samples characterized by P binary features}

\item{Y}{An N x 1 binary vector, which is the binarized version of the continuous output variable}

\item{W}{An N x 1 continuous vector with weights for each of the N samples}

\item{K}{The number of disjunctive terms}

\item{M}{The maximum number of selected features per disjunctive term}

\item{lambda}{The regularizer of penalty for model complexity}

\item{sens}{The constraints on minimum sensitivity}

\item{spec}{The constraints on minimum specificity}

\item{addcons}{Some additional constraints}
}
\value{
What ever this function returns
}
\description{
An R helper function implemented as a wrapper for underlying C code
}
