% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF_CPLEX_weak_pos.R
\name{DNF_CPLEX_weak_pos}
\alias{DNF_CPLEX_weak_pos}
\title{Compute DNF CPLEX weak pos}
\usage{
DNF_CPLEX_weak_pos(X, Y, W, K, M, lambda, sens, spec, addcons)
}
\arguments{
\item{X}{A matrix}

\item{Y}{A binary matrix of logical cateogorizations}

\item{W}{Some parameters}

\item{K}{Some parameters}

\item{M}{Some parameters}

\item{lambda}{A parameter}

\item{sens}{A parameter}

\item{spec}{Some more parmaters}

\item{addcons}{Some stuff}
}
\description{
A function that excecutes some part of a logical model that I don't know about
}
\examples{
test <- 2 +2
test

}
