% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF_ILP_weak.R
\name{DNF_ILP_weak}
\alias{DNF_ILP_weak}
\title{Compute DNF CPLEX weak pos}
\usage{
DNF_ILP_weak(X, Y, W, K, M, lambda, sens, spec, addcons)
}
\arguments{
\item{X}{An N x P binary matrix with N samples characterized by P binary features}

\item{Y}{An N x 1 binary vector, which is the binarized version of the continuous output variable}

\item{W}{An N x 1 continuous vector with weights for each of the N samples}

\item{K}{The number of disjunctive terms}

\item{M}{The maximum number of selected features per disjunctive term}

\item{lambda}{The regularizer of penalty for model complexity}

\item{sens}{The constraints on minimum sensitivity}

\item{spec}{The constraints on minimum specificity}

\item{addcons}{Some additional constraints}
}
\value{
What ever this function returns
}
\description{
A function that excecutes some part of a logical model that I don't know about
}
\examples{
test <- 2 +2
test

}
