% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsolution.R
\name{.getsolution}
\alias{.getsolution}
\title{Get Lobico Model Solution}
\usage{
.getsolution(x, K, M, P)
}
\arguments{
\item{x}{The  matrix}

\item{K}{The number of disjunctive terms}

\item{M}{The maximum number of selected features per disjunctive term}

\item{P}{The number of rows of solution matrix}
}
\value{
The solution matrix to the lobico model
}
\description{
Input model parameters and output result
}
