% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showformula.R
\name{.showformula}
\alias{.showformula}
\title{Show the Lobico Formula}
\usage{
.showformula(SolMat, K, M, X1)
}
\arguments{
\item{X1}{}
}
\description{
Displays the lobico formula used in the logical model
}
