% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobico.R
\name{lobico}
\alias{lobico}
\title{Lobico help function}
\usage{
lobico(X, Y, K, M, spec, sens, lambda, weak, pos, addcons)
}
\arguments{
\item{X}{An N x P binary matrix with N samples characterized by P binary features}

\item{Y}{An N x 1 continuous vector with weights for each of the N samples}

\item{K}{The number of disjunctive terms}

\item{M}{The maximum number of selected features per disjunctive term}

\item{spec}{The constraints on minimum specificity}

\item{sens}{The constraints on minimum sensitivity}

\item{lambda}{The regularizer of penalty for model complexity}

\item{weak}{The parameter about whether to formulate weaker logic model}

\item{pos}{The parameter about whether to return quick and dirty solutions to disallow negations, i.e. only positive}

\item{addcons}{Some additional constraints}
}
\value{

}
\description{
A wrapper for implementing C functions for calculating logical models
}
\examples{
data(bibw2992)


}
