% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNF_CPLEX_weak_pos.R
\name{CNF_CPLEX_weak_pos}
\alias{CNF_CPLEX_weak_pos}
\title{CNF CPLEX weak position function}
\usage{
CNF_CPLEX_weak_pos(X, Y, W, K, M, lambda, sens, spec, addcons)
}
\arguments{
\item{X}{A N x P binary matrix with N samples characterized by P binary features}

\item{Y}{A N x 1 binary vector, which is the binarized version of the continuous output variable}

\item{W}{A N x 1 continous vector with weights for each version of the continuous output variable}

\item{K}{A model parameter}

\item{M}{A model parameter}

\item{lambda}{A model parameter}

\item{sens}{The minimum sensitivity}

\item{spec}{The minimum specificity}

\item{addcons}{Some additional parameters}
}
\value{
What ever this function returns
}
\description{
An R helper function implemented as a wrapper for underlying C code
}
\examples{
test <- 2 +2
test

}
