% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsolution.R
\name{.getsolution}
\alias{.getsolution}
\title{Get Lobico Model Solution}
\usage{
.getsolution(x, K, M, P)
}
\arguments{
\item{x}{Parmeters}

\item{K}{Parameters}

\item{M}{Parameters}

\item{P}{Parameters}
}
\description{
Input model parameters and output result
}
