% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showformula.R
\name{.showformula}
\alias{.showformula}
\title{Show the Lobico Formula}
\usage{
.showformula(SolMat, K, M, X1)
}
\arguments{
\item{SolMat}{The solution matrix}

\item{K}{The number of disjunctive terms}

\item{M}{The maximum number of selected features per disjunctive term}

\item{X1}{The features matrix}
}
\value{
The string represents the solution
}
\description{
Displays the lobico formula used in the logical model
}
