% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobico.R
\name{lobico}
\alias{lobico}
\title{Lobico help function}
\usage{
lobico(X, Y, K, M, solve, param, spec, sens, lambda, weak, pos, addcons)
}
\arguments{
\item{X}{A data matrix}

\item{Y}{A binary matrix of logical categorizations}

\item{K}{A parameter}

\item{M}{A parameter}

\item{solve}{A parameter}

\item{param}{A parameter}

\item{spec}{A parameter}

\item{sens}{A parameter}

\item{lambda}{A parameter}

\item{weak}{A parameter}

\item{pos}{A parameter}

\item{addcons}{Some other stuff}
}
\value{
What ever this function returns
}
\description{
A wrapper for implementing C functions for calculating logical models
}
\examples{
test <- (2 + 2)
test

}
